+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # AirDash Changelog
 
+## Current version (in development)
+
+* Feature: Support reading AIS data from serial port.
+* Breaking change: URLs for AIS TCP now have the protocol `ais-tcp://`
+
 ## v0.2.0 (2021-01-01)
 
 * Data source configuration through `DATA_SOURCES` env.
 * All data source connectivity is now managed server-side.
 * Track logs are now visible.
+
 ## v0.1.0 (2020-12-28)
 
 * First release.